Corrige les messages d'erreurs pour qu'ils ne fassent plus crash le programme. Ajoute un peu de couleur dans l'affichage

+++ tp1/TestDev/training.py
@@ -12,7 +12,7 @@ class Training():
 
                 self.dict = self.create_dict(self.words)
                 self.matrix = self._create_matrix(self.dict)
-                self.fill_matrix_vector(self.matrix, self.words, self.dict, window)
+                self.fill_matrix_vector(self.matrix, self.words, self.dict, (int) window)
                 self.lookup_table = self.create_lookup_table()
 
         def create_lookup_table(self):
